Dev Merge changes (#27)
* - Due to mobile view issues, revert back to modal nav menu instead of menu at top of page
- Go back to previous Nav menu design as it works better
- Change Welcome display to time and date display to be more descriptive
- Add a component called welcome display as a stub to welcome logged in users with no tasks
- Add a component called not logged in as a stub to let users know they are not logged in on home page
- Rename No Task display to No Active Task display to be more descriptive
- Add new component called No Task display when there is no tasks and the user may or may not be logged in

* - Remove close icon from dialog modal(will readd for needed modals in their own files)
 - Re-fixed links on nav menu having click deadzones
 - Slightly adjust color of logo to lighter green
 - Add TODO comment in login to re-add close button in that file
 - Redesign of Task Item details layout and form to accomodate both larger and smaller screens
 - Add back cancel buttons to modals since close icons are gone again
 - Adjust button aesthetics to be visually consistent across the app
 - Send user to home page if they logout instead of staying on current page
 - Create Welcome Landing page to let user quickly add task if no tasks are created yet.
 - Created Not logged in page to tell user to log in

* Remove the remnants of close button from Dialog class
Fix Nav Menu size on larger screens
Fix Task details modal on larger screens
Create function to close modal in Login and use it everywhere in logint hat close the modal

* - Lots of responsive changes for larger screens and smallers screens across app
- Add status checking for every task query to check if task was actually completed correctly and notify user via toast notification to let user know what is happening
- Add useEfffect for add task and unassign/complete task that checks context to show toast after screen refreshes with new info. To be added to edit and delete later on
- Renaming for clearer and more descriptive components
- Revamp visual task list for assign list to be compact and hide unncessary information.
- Add user ID check for all task query operations to be extra safe.
- Remove gray backgrounds and change task item cards to be gray instead for more consistent look across app.
- On large screens, permanently show side menu. Look is a WIP
- Break down complete Task query into recurring and normal complete task query. Further break down each query in the recurring and normal into their own simpler queries.
- Remove Modal background blurring

* - Remove Status checking in every query as it did not work as intended to check if something went wrong.
- Use Toast Promise to show loading notifications and success or error notifications.
- Revamp task Item action buttons to now always be visible for manage tasks to edit or delete.
- With refactor, task item actions no longer needed
- Remove useeffects for toasts since the toast promise calls completely replace this with added benefits of being where thet action is completed
- More renaming to make things clearer
- Remove use of Context with the toast promises
- Add required to the calendar so you can't unselect the date

* - Fix date calculation for Next due date to correct set the due date to the next day from the due date and not current date
- Fixed Querykeys to be more descriptive and properly reset them when Updates happen
- Fix bug where clicking on any task item in Find Task would only default to one specific task Item
- Experiment to show loading screens in Initial Task Display. Still a WIP
- New view for Active Task Details to hide unnecessary information
- Update all outdated NPM packages using npm-check-updates

* Fix messed up loading of home page where it would show no task screen first then show the correct screen by adding an intermediate screen to load it properly

* - Customize calendar width, height and font size
- Allow full width of elements in NavBar children
- Remove hardcoded width limits for larger screens
- Refactor CSS for calendar
- Move priority icon to start of line

+++ components/CRUD_queries/DeleteTaskQuery.tsx
@@ -1,14 +1,17 @@
 import supabaseClient from "@/lib/supabaseClient";
 
-const DeleteTaskQuery = async (taskId: string) => {
-  const { error: supabaseError } = await supabaseClient
+const DeleteTaskQuery = async (taskId: string, userId: string) => {
+  const { status, error } = await supabaseClient
     .from("tasks")
     .delete()
-    .eq("id", taskId);
+    .eq("id", taskId)
+    .eq("created_by", userId);
 
-  if (supabaseError) {
-    throw new Error(supabaseError.message);
-  }
+  if (error) throw new Error(error.message);
+
+  if (status === 204) return true;
+
+  return false;
 };
 
 export default DeleteTaskQuery;
